♻️refactor: unify line related functions
