+++ app/area/hiroshima/page.tsx
@@ -1,8 +1,8 @@
 import { Register } from "@/app/area/register"
 import { Schedules } from "@/app/area/schedules"
+import { LineRegister } from "@/app/components/button/lineAddFriends"
 import { ScheduleCarousel } from "@/app/components/layout/carousel"
 import { Heading } from "@/app/components/layout/heading"
-import { LineRegister } from "@/app/components/layout/lineRegister"
 import { Navigation } from "@/app/components/layout/navigation"
 import type { Guideline } from "@/app/interfaces/guideline"
 import type { Schedule } from "@/app/interfaces/schedule"
